# docs/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Legal Metrology Compliance Scanner",
    "description": "Backend API for Legal Metrology Compliance Scanner (Smart India Hackathon 2026)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/dashboard/trends": {
      "get": {
        "tags": [
          "Executive Analytics & Dashboard"
        ],
        "summary": "Get Time-Series Compliance Trends",
        "description": "Returns daily inspection volumes, compliant counts, and compliance rates.",
        "operationId": "get_dashboard_trends_api_v1_dashboard_trends_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Number of historical days to analyze",
              "default": 30,
              "title": "Days"
            },
            "description": "Number of historical days to analyze"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrendItemSchema"
                  },
                  "title": "Response Get Dashboard Trends Api V1 Dashboard Trends Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/top-violations": {
      "get": {
        "tags": [
          "Executive Analytics & Dashboard"
        ],
        "summary": "Get Top Statutory Violations",
        "description": "Returns most frequently breached rules under Legal Metrology PCR 2011.",
        "operationId": "get_top_violations_api_v1_dashboard_top_violations_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Max violations to return",
              "default": 5,
              "title": "Limit"
            },
            "description": "Max violations to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopViolationItemSchema"
                  },
                  "title": "Response Get Top Violations Api V1 Dashboard Top Violations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/jurisdiction-heatmap": {
      "get": {
        "tags": [
          "Executive Analytics & Dashboard"
        ],
        "summary": "Get GIS Jurisdiction Compliance Heatmap",
        "description": "Computes district and state-level compliance statistics and average GPS centroids for GIS mapping.",
        "operationId": "get_jurisdiction_heatmap_api_v1_dashboard_jurisdiction_heatmap_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JurisdictionHeatmapItemSchema"
                  },
                  "type": "array",
                  "title": "Response Get Jurisdiction Heatmap Api V1 Dashboard Jurisdiction Heatmap Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/dashboard/repeat-offenders": {
      "get": {
        "tags": [
          "Executive Analytics & Dashboard"
        ],
        "summary": "Get Repeat Offenders Leaderboard",
        "description": "Tracks manufacturers and brands with highest repeat infraction frequency.",
        "operationId": "get_repeat_offenders_api_v1_dashboard_repeat_offenders_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max offenders to list",
              "default": 10,
              "title": "Limit"
            },
            "description": "Max offenders to list"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepeatOffenderLeaderboardItemSchema"
                  },
                  "title": "Response Get Repeat Offenders Api V1 Dashboard Repeat Offenders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JurisdictionHeatmapItemSchema": {
        "properties": {
          "district": {
            "type": "string",
            "title": "District"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "total_inspections": {
            "type": "integer",
            "title": "Total Inspections"
          },
          "compliant_count": {
            "type": "integer",
            "title": "Compliant Count"
          },
          "non_compliant_count": {
            "type": "integer",
            "title": "Non Compliant Count"
          },
          "compliance_rate": {
            "type": "number",
            "title": "Compliance Rate"
          },
          "centroid_latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Centroid Latitude"
          },
          "centroid_longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Centroid Longitude"
          }
        },
        "type": "object",
        "required": [
          "district",
          "state",
          "total_inspections",
          "compliant_count",
          "non_compliant_count",
          "compliance_rate"
        ],
        "title": "JurisdictionHeatmapItemSchema"
      },
...
      "RepeatOffenderLeaderboardItemSchema": {
        "properties": {
          "manufacturer_name": {
            "type": "string",
            "title": "Manufacturer Name"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "total_violations": {
            "type": "integer",
            "title": "Total Violations"
          },
          "inspections_count": {
            "type": "integer",
            "title": "Inspections Count"
          },
          "last_inspection_date": {
            "type": "string",
            "title": "Last Inspection Date"
          }
        },
        "type": "object",
        "required": [
          "manufacturer_name",
          "brand_name",
          "total_violations",
          "inspections_count",
          "last_inspection_date"
        ],
        "title": "RepeatOffenderLeaderboardItemSchema"
      },
...
      "TopViolationItemSchema": {
        "properties": {
          "rule_code": {
            "type": "string",
            "title": "Rule Code"
          },
          "section_violated": {
            "type": "string",
            "title": "Section Violated"
          },
          "violation_title": {
            "type": "string",
            "title": "Violation Title"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "violation_count": {
            "type": "integer",
            "title": "Violation Count"
          },
          "percentage_of_total": {
            "type": "number",
            "title": "Percentage Of Total"
          }
        },
        "type": "object",
        "required": [
          "rule_code",
          "section_violated",
          "violation_title",
          "severity",
          "violation_count",
          "percentage_of_total"
        ],
        "title": "TopViolationItemSchema"
      },
      "TrendItemSchema": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "total_inspections": {
            "type": "integer",
            "title": "Total Inspections"
          },
          "compliant_count": {
            "type": "integer",
            "title": "Compliant Count"
          },
          "non_compliant_count": {
            "type": "integer",
            "title": "Non Compliant Count"
          },
          "compliance_rate": {
            "type": "number",
            "title": "Compliance Rate"
          }
        },
        "type": "object",
        "required": [
          "date",
          "total_inspections",
          "compliant_count",
          "non_compliant_count",
          "compliance_rate"
        ],
        "title": "TrendItemSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}